+++ vigi/app.py
@@ -4,6 +4,7 @@ from flask_bootstrap import Bootstrap5
 from flask import Flask, redirect, url_for
 
 from .cache import cache
+from .auth import auth
 from .context_processors import utility_processor
 
 from .routes.live import live_blueprint
@@ -26,6 +27,12 @@ bootstrap = Bootstrap5(app)
 app.register_blueprint(live_blueprint)
 app.register_blueprint(camera_blueprint)
 app.register_blueprint(recordings_blueprint)
+
+# configure the basic auth for all routes
+@app.before_request
+@auth.login_required
+def before_request():
+    pass
 
 # the root route redirects to the live view
 @app.route('/')